COVER-43: Renamed search index commands for consistency
