+++ src/Command/Index/IndexPopulateCommand.php
@@ -4,8 +4,9 @@
  * Contains a command to populate the search index.
  */
 
-namespace App\Command;
+namespace App\Command\Index;
 
+use App\Exception\SearchIndexException;
 use App\Service\PopulateService;
 use App\Service\VendorService\ProgressBarTrait;
 use Symfony\Component\Console\Attribute\AsCommand;
@@ -16,10 +17,13 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * Class SearchPopulateCommand.
+ * Class IndexPopulateCommand.
  */
-#[AsCommand(name: 'app:search:populate')]
-class SearchPopulateCommand extends Command
+#[AsCommand(
+    name: 'app:index:populate',
+    description: 'Populate the search index with data from the search table.'
+)]
+class IndexPopulateCommand extends Command
 {
     use ProgressBarTrait;
 
@@ -40,7 +44,6 @@ class SearchPopulateCommand extends Command
     protected function configure(): void
     {
         $this
-            ->setDescription('Populate the search index with data from the search table.')
             ->addOption('force', null, InputOption::VALUE_NONE, 'Force execution ignoring locks')
             ->addOption('id', null, InputOption::VALUE_OPTIONAL, 'Single search table record id (try populate single record)', -1);
     }
@@ -50,6 +53,7 @@ class SearchPopulateCommand extends Command
      *
      * @throws \Doctrine\ORM\NoResultException
      * @throws \Doctrine\ORM\NonUniqueResultException
+     * @throws SearchIndexException
      */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
